added first test case
