+++ ghostdriver.log
@@ -1,5 +1,5 @@
-[INFO  - 2017-10-17T21:31:19.221Z] GhostDriver - Main - running on port 65403
-[INFO  - 2017-10-17T21:31:21.402Z] Session [84d85560-b382-11e7-b597-51425ffabef8] - page.settings - {"XSSAuditingEnabled":false,"javascriptCanCloseWindows":true,"javascriptCanOpenWindows":true,"javascriptEnabled":true,"loadImages":true,"localToRemoteUrlAccessEnabled":false,"userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X) AppleWebKit/538.1 (KHTML, like Gecko) PhantomJS/2.1.1 Safari/538.1","webSecurityEnabled":true}
-[INFO  - 2017-10-17T21:31:21.402Z] Session [84d85560-b382-11e7-b597-51425ffabef8] - page.customHeaders:  - {}
-[INFO  - 2017-10-17T21:31:21.402Z] Session [84d85560-b382-11e7-b597-51425ffabef8] - Session.negotiatedCapabilities - {"browserName":"phantomjs","version":"2.1.1","driverName":"ghostdriver","driverVersion":"1.2.0","platform":"mac-unknown-64bit","javascriptEnabled":true,"takesScreenshot":true,"handlesAlerts":false,"databaseEnabled":false,"locationContextEnabled":false,"applicationCacheEnabled":false,"browserConnectionEnabled":false,"cssSelectorsEnabled":true,"webStorageEnabled":false,"rotatable":false,"acceptSslCerts":false,"nativeEvents":true,"proxy":{"proxyType":"direct"}}
-[INFO  - 2017-10-17T21:31:21.402Z] SessionManagerReqHand - _postNewSessionCommand - New Session Created: 84d85560-b382-11e7-b597-51425ffabef8
+[INFO  - 2017-10-17T22:48:22.716Z] GhostDriver - Main - running on port 51298
+[INFO  - 2017-10-17T22:48:25.023Z] Session [48bce8b0-b38d-11e7-bb0d-b7b7457ab108] - page.settings - {"XSSAuditingEnabled":false,"javascriptCanCloseWindows":true,"javascriptCanOpenWindows":true,"javascriptEnabled":true,"loadImages":true,"localToRemoteUrlAccessEnabled":false,"userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X) AppleWebKit/538.1 (KHTML, like Gecko) PhantomJS/2.1.1 Safari/538.1","webSecurityEnabled":true}
+[INFO  - 2017-10-17T22:48:25.023Z] Session [48bce8b0-b38d-11e7-bb0d-b7b7457ab108] - page.customHeaders:  - {}
+[INFO  - 2017-10-17T22:48:25.023Z] Session [48bce8b0-b38d-11e7-bb0d-b7b7457ab108] - Session.negotiatedCapabilities - {"browserName":"phantomjs","version":"2.1.1","driverName":"ghostdriver","driverVersion":"1.2.0","platform":"mac-unknown-64bit","javascriptEnabled":true,"takesScreenshot":true,"handlesAlerts":false,"databaseEnabled":false,"locationContextEnabled":false,"applicationCacheEnabled":false,"browserConnectionEnabled":false,"cssSelectorsEnabled":true,"webStorageEnabled":false,"rotatable":false,"acceptSslCerts":false,"nativeEvents":true,"proxy":{"proxyType":"direct"}}
+[INFO  - 2017-10-17T22:48:25.023Z] SessionManagerReqHand - _postNewSessionCommand - New Session Created: 48bce8b0-b38d-11e7-bb0d-b7b7457ab108
